exp: finish leaf wrap 512 and 1024 test

+++ include/psi/dependence/tree_node.h
@@ -35,7 +35,7 @@ struct Node {
   size_t size;
 };
 
-template <typename Point, typename Range, uint_fast8_t kDefaultWrap,
+template <typename Point, typename Range, uint_fast16_t kDefaultWrap,
           typename AugType = std::monostate,
           typename PointAssignTag = parlay::move_assign_tag>
 struct LeafNode : Node {
